+++ Changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+ - Sketchpad: Small gap at the bottom of view mode image
+
 ### Added
  - Tagpad: Automatic resizing of drawing area to fill half of available width